app/backend/services/auth-service/requirements.txt
# FastAPI та веб-фреймворк
fastapi==0.104.1
uvicorn==0.24.0

# База даних
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1

# JWT та безпека
PyJWT==2.8.0
bcrypt==4.1.2
cryptography==41.0.7

# MFA
pyotp==2.9.0

# HTTP клієнт
requests==2.31.0

# Налаштування
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Логування
loguru==0.7.2

# Тестування
pytest==7.4.3
pytest-asyncio==0.21.1 